// backoffice/src/models/Product.ts
export type Product = {
  id: number;
  name: string;
  price: number;
  isDiscount: boolean;
  discountPercent: number;
  discountPrice: number;
  image: string;
  slug: string;
  description: string;
  category: string;
  ratingStar: number;
  ratingCount: number;
  soldCount: number;
  inStockCount: number;
};